+++ src/main.rs
@@ -1,13 +1,13 @@
+/* Copyright © 2025, CosmicMind, Inc. */
+
 use entid::{Prefix, UuidEntityId};
 
-// Use the derive macro to implement the Prefix trait
-#[derive(Debug, Prefix)]
-#[prefix = "user"]
+#[derive(Prefix)]
+#[entid(prefix = "user", delimiter = "_")]
 struct User;
 
 fn main() {
-    // Now we can use the User struct with EntityId
     let user_id = UuidEntityId::<User>::generate();
     println!("User ID: {}", user_id);
-    assert!(user_id.as_str().starts_with("user_"));
+    assert!(user_id.to_string().starts_with("user_"));
 }